pyproject.toml
[project]
name = "home-monitoring"
version = "0.1.0"
description = "Home monitoring system"
requires-python = ">=3.12.3"
dependencies = [
    "aioinflux==0.9.0",
    "fastapi==0.110.0",
    "httpx==0.27.0",
    "pandas==2.0.3",
    "pydantic==2.6.1",
    "pydantic-settings==2.1.0",
    "structlog==24.1.0",
    "lnetatmo>=4.2.0",
    "solaredge==0.0.4",
    "py-smart-gardena==1.3.9",
    "pyTibber==0.32.2",
    "pyserial==3.5",
    "matplotlib==3.8.2",
    "oauthlib==3.2.2",
    "setuptools>=69.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.2",
    "pytest-asyncio==0.23.5",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0",
    "black==24.2.0",
    "ruff==0.2.1",
    "mypy==1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/home_monitoring"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
